harryprt/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.9
botocore==1.42.9
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
faiss-cpu==1.13.1
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.12.0
gradio==6.1.0
gradio-client==2.0.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.2
langchain-aws==1.1.0
langchain-core==1.1.1
langchain-huggingface==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
s3transfer==0.16.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
tenacity==9.1.2
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
